api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.123.4",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "pwdlib[argon2]>=0.3.0",
    "asyncpg>=0.31.0",
    "boto3>=1.35.0",
    "alembic>=1.17.2",
    "shortuuid>=1.0.13",
]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
